Split parser declarations into separate modules
