+++ Makefile
@@ -6,7 +6,8 @@ BIN = vc
 # Core compiler sources
 
 CORE_SRC = src/main.c src/compile.c src/compile_stage.c src/compile_link.c src/compile_tokenize.c src/compile_parse.c src/compile_output.c src/startup.c src/command.c src/cli.c src/cli_env.c src/cli_opts.c src/lexer.c src/ast_expr.c src/ast_stmt.c src/ast_clone.c src/parser_core.c src/parser_toplevel.c src/symtable_core.c src/symtable_globals.c src/symtable_struct.c src/parser_expr.c src/parser_expr_primary.c src/parser_expr_binary.c src/parser_init.c \
-           src/parser_decl.c src/parser_flow.c src/parser_stmt.c src/parser_types.c \
+           src/parser_decl_var.c src/parser_decl_struct.c src/parser_decl_enum.c \
+           src/parser_flow.c src/parser_stmt.c src/parser_types.c \
            src/semantic_expr.c src/semantic_arith.c src/semantic_mem.c src/semantic_call.c \
            src/semantic_loops.c src/semantic_switch.c src/semantic_init.c src/semantic_var.c src/semantic_stmt.c src/semantic_inline.c src/semantic_global.c src/consteval.c src/error.c src/ir_core.c src/ir_const.c src/ir_memory.c src/ir_control.c src/ir_global.c \
            src/codegen.c src/codegen_mem.c src/codegen_loadstore.c src/codegen_arith_int.c src/codegen_arith_float.c src/codegen_branch.c \
@@ -119,9 +120,15 @@ src/parser_expr_binary.o: src/parser_expr_binary.c $(HDR)
 src/parser_init.o: src/parser_init.c $(HDR)
 	$(CC) $(CFLAGS) $(OPTFLAGS) -Iinclude -c src/parser_init.c -o src/parser_init.o
 
-src/parser_decl.o: src/parser_decl.c $(HDR)
-	$(CC) $(CFLAGS) $(OPTFLAGS) -Iinclude -c src/parser_decl.c -o src/parser_decl.o
-
+src/parser_decl_var.o: src/parser_decl_var.c $(HDR)
+	$(CC) $(CFLAGS) $(OPTFLAGS) -Iinclude -c src/parser_decl_var.c -o src/parser_decl_var.o
+	
+src/parser_decl_struct.o: src/parser_decl_struct.c $(HDR)
+	$(CC) $(CFLAGS) $(OPTFLAGS) -Iinclude -c src/parser_decl_struct.c -o src/parser_decl_struct.o
+	
+src/parser_decl_enum.o: src/parser_decl_enum.c $(HDR)
+	$(CC) $(CFLAGS) $(OPTFLAGS) -Iinclude -c src/parser_decl_enum.c -o src/parser_decl_enum.o
+	
 src/parser_flow.o: src/parser_flow.c $(HDR)
 	$(CC) $(CFLAGS) $(OPTFLAGS) -Iinclude -c src/parser_flow.c -o src/parser_flow.o
