+++ include/parser_decl_enum.h
@@ -1,0 +1,8 @@
+#ifndef VC_PARSER_DECL_ENUM_H
+#define VC_PARSER_DECL_ENUM_H
+
+#include "parser.h"
+
+stmt_t *parser_parse_enum_decl(parser_t *p);
+
+#endif /* VC_PARSER_DECL_ENUM_H */
