+++ src/parser_toplevel.c
@@ -11,6 +11,9 @@
 #include "parser_types.h"
 #include "ast_stmt.h"
 #include "ast_expr.h"
+#include "parser_decl_var.h"
+#include "parser_decl_struct.h"
+#include "parser_decl_enum.h"
 #include "symtable.h"
 #include "util.h"
 #include "error.h"
